addressing other PR feedback
